restructure app directories
