+++ functions/find_gif/gifs.json
@@ -53,15 +53,15 @@
 }, {
   "URL": "https://media0.giphy.com/media/2VUIE0b9l1Sdq/giphy.gif",
   "alt_text": "Pleaceful plants in a pastel picture",
-  "tags": ["vibes"]
+  "tags": ["plants"]
 }, {
   "URL": "https://media3.giphy.com/media/IbsQ8L4PTQ53SH7FnL/giphy.gif",
   "alt_text": "A potted plant receives much needed affection",
-  "tags": ["vibes"]
+  "tags": ["plants"]
 }, {
   "URL": "https://media1.giphy.com/media/xUA7aOIFDR4ZgqLy8w/giphy.gif",
   "alt_text": "Fern having a messy hair day",
-  "tags": ["vibes"]
+  "tags": ["plants"]
 }, {
   "URL": "https://media1.giphy.com/media/MbAlP79yMRysHKUyHV/giphy.gif",
   "alt_text": "Sleepy otter rubs checks and yawns",
