--- functions/index.ts
@@ -1,3 +1,0 @@
-import FindGIF from "./find-gif/definition.ts";
-
-export { FindGIF };
